+++ modules/@apostrophecms/login/index.js
@@ -67,7 +67,9 @@ module.exports = {
       allowedAttempts: 3,
       perMinutes: 1,
       lockoutMinutes: 1
-    }
+    },
+    minimumWhoamiFields: [ '_id', 'username', 'title', 'email' ],
+    whoamiFields: []
   },
   async init(self) {
     self.passport = new Passport();
@@ -346,6 +348,28 @@ module.exports = {
         // it should be accessed via POST because the result
         // may differ by individual user session and should not
         // be cached
+        async whoami (req) {
+          if (!req.user) {
+            throw self.apos.error('notfound');
+          }
+
+          const allowedFields = new Set([ 'role' ]);
+
+          // The query will contain comma separated additional field names
+          const requestedFields = req.query.additionalFields
+            ? req.query.additionalFields.split(',').map(field => field.trim()).filter(field => allowedFields.has(field)) : [];
+
+          const fields = new Set([ ...self.options.minimumWhoamiFields, ...self.options.whoamiFields, ...requestedFields ]);
+          const user = {};
+
+          for (const field of fields) {
+            if (req.user[field] !== undefined) {
+              user[field] = req.user[field];
+            }
+          }
+
+          return user;
+        },
         async context(req) {
           return self.getContext(req);
         },